requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
Django==3.1.3
django-extensions==3.1.0
django-filter==2.4.0
djangorestframework==3.12.2
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==1.0.0
Pillow==8.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.0
python-ldap @ file:///C:/Users/Joel/Documents/Proyectos/ChatREST/python_ldap-3.3.1-cp39-cp39-win_amd64.whl
pytz==2020.4
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted @ file:///C:/Users/Joel/Documents/Proyectos/ChatREST/Twisted-20.3.0-cp39-cp39-win_amd64.whl
txaio==20.4.1
zope.interface==5.2.0
